edited code

+++ src/assignment2/Pokemon.java
@@ -119,7 +119,7 @@ public class Pokemon {
         this.hasFainted = true;
     }
 
-    public void receiveDamage(int damageValue) {
+    public void receiveDamage(double damageValue) {
         this.currentHP -= damageValue;
         if (this.currentHP <= 0) {
             this.currentHP = 0;
